sixth commit
the main part of this commit would be  editing some user profile and upload the real video and connect the video and user in the database so that we can know which video is belongs to which user.
second part is done with webpack which is a skeleton part before doing scss.

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/video";
+import User from "../models/User";
 
 export const home = async (req, res) => {
 	const videos = await Video.find({});
@@ -6,7 +7,7 @@ export const home = async (req, res) => {
 };
 export const watch = async (req, res) => {
 	const { id } = req.params;
-	const video = await Video.findById(id);
+	const video = await Video.findById(id).populate("owner");
 	if (!video) {
 		return res.status(404).render("404", { pageTitle: "VIDEO NOT FOUND." });
 	}
@@ -15,20 +16,34 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
 	const { id } = req.params;
+	const {
+		user: { _id },
+	} = req.session;
 	const video = await Video.findById(id);
 	if (!video) {
 		return res.status(404).render("404", { pageTitle: "VIDEO NOT FOUND." });
+	}
+	if (String(video.owner) !== String(_id)) {
+		return res.status(403).redirect("/");
 	}
 	return res.render("edit", { pageTitle: `edit ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
 	const { id } = req.params;
+	const {
+		user: { _id },
+	} = req.session;
 	const { title, description, hashtags } = req.body;
 	const video = await Video.exists({ _id: id });
 	if (!video) {
 		return res.status(404).render("404", { pageTitle: "VIDEO NOT FOUND." });
 	}
+
+	if (String(video.owner) !== String(_id)) {
+		return res.status(403).redirect("/");
+	}
+
 	await Video.findByIdAndUpdate(id, {
 		title,
 		description,
@@ -43,17 +58,26 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+	const {
+		user: { _id },
+	} = req.session;
+	const { title, description, hashtags } = req.body;
+	const file = req.file;
 	try {
-		const { title, description, hashtags } = req.body;
-		await Video.create({
+		const newVideo = await Video.create({
 			title,
 			description,
+			videoURL: file.path,
+			owner: _id,
 			hashtags: Video.formatHashtags(hashtags),
 		});
+		const user = await User.findById(_id);
+		user.videos.push(newVideo._id);
+		user.save();
 		return res.redirect("/");
 	} catch (error) {
 		return res.status(400).render("upload", {
-			pageTitle: `Upload video`,
+			pageTitle: `Upload failed`,
 			errorMsg: error._message,
 		});
 	}
@@ -61,6 +85,16 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
 	const { id } = req.params;
+	const {
+		user: { _id },
+	} = req.session;
+	const video = await Video.findById(id);
+	if (!video) {
+		return res.status(404).render("404", { pageTitle: "VIDEO NOT FOUND." });
+	}
+	if (String(video.owner) !== String(_id)) {
+		return res.status(403).redirect("/");
+	}
 	await Video.findByIdAndDelete(id);
 	return res.redirect("/");
 };
